+++ modules/services/media-noop/src/main/kotlin/au/com/shiftyjelly/pocketcasts/media/NoOpMediaService.kt
@@ -1,6 +1,5 @@
 package au.com.shiftyjelly.pocketcasts.media
 
-import android.content.Context
 import au.com.shiftyjelly.pocketcasts.models.entity.Podcast
 import au.com.shiftyjelly.pocketcasts.models.entity.PodcastEpisode
 import au.com.shiftyjelly.pocketcasts.sharing.Clip
@@ -8,9 +7,7 @@ import au.com.shiftyjelly.pocketcasts.sharing.MediaService
 import au.com.shiftyjelly.pocketcasts.sharing.VisualCardType
 import java.io.File
 
-internal class NoOpMediaService(
-    private val context: Context,
-) : MediaService {
+internal class NoOpMediaService : MediaService {
     override suspend fun clipAudio(podcast: Podcast, episode: PodcastEpisode, clipRange: Clip.Range) = runCatching {
         error("Operation not supported")
     }